+++ src/main/scala/ee/hrzn/chryse/platform/resource/ClockSource.scala
@@ -20,8 +20,8 @@ package ee.hrzn.chryse.platform.resource
 
 import chisel3._
 
-case class ClockSource(hz: Int) extends ResourceData(Input(Clock())) {}
+case class ClockSource(hz: BigInt) extends ResourceData(Input(Clock())) {}
 
 object ClockSource {
-  def apply(hz: Int) = new ClockSource(hz)
+  def apply(hz: BigInt) = new ClockSource(hz)
 }
